Спринт 10. Домашнее задание. Версия для ревью 2.0
