+++ app/src/main/java/com/practicum/playlistmaker/Track.kt
@@ -1,0 +1,8 @@
+package com.practicum.playlistmaker
+
+data class Track(
+    val trackName: String,
+    val artistName: String, // Имя исполнителя
+    val trackTime: String, // Продолжительность трека
+    val artworkUrl100: String // Ссылка на изображение обложки
+)
