+++ app/src/main/java/com/practicum/playlistmaker/TrackListComponents.kt
@@ -1,0 +1,49 @@
+package com.practicum.playlistmaker
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.ImageView
+import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
+import com.bumptech.glide.Glide
+import com.bumptech.glide.load.engine.DiskCacheStrategy
+import com.bumptech.glide.load.resource.bitmap.RoundedCorners
+
+class TrackViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+    private val trackName: TextView = itemView.findViewById(R.id.trackName)
+    private val artistName: TextView = itemView.findViewById(R.id.artistName)
+    private val image = itemView.findViewById<ImageView>(R.id.image)
+
+    fun bind(track: Track) {
+        trackName.text = track.trackName
+        artistName.text = track.artistName + " • " + track.trackTime
+
+        val radiusInPixels = itemView.context.getResources()
+            .getDimensionPixelSize(R.dimen.tracklst_item_image_radius)
+
+        Glide.with(itemView)
+            .load(track.artworkUrl100)
+            .diskCacheStrategy(DiskCacheStrategy.NONE)
+            .skipMemoryCache(true)
+            .placeholder(R.drawable.placeholder)
+            .centerCrop()
+            .transform(RoundedCorners(radiusInPixels))
+            .into(image)
+    }
+}
+
+class TrackAdapter(private val tracks: List<Track>) : RecyclerView.Adapter<TrackViewHolder>() {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): TrackViewHolder {
+        val view = LayoutInflater.from(parent.context).inflate(R.layout.track_view, parent, false)
+        return TrackViewHolder(view)
+    }
+
+    override fun onBindViewHolder(holder: TrackViewHolder, position: Int) {
+        holder.bind(tracks[position])
+    }
+
+    override fun getItemCount(): Int {
+        return tracks.size
+    }
+}
