+++ app/src/main/res/layout/track_view.xml
@@ -34,34 +34,29 @@
         <TextView
             android:id="@+id/trackName"
 
+            style="@style/TrackViewFirstLineStyle"
+
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
 
-            android:fontFamily="@font/ys_display_regular"
             android:maxLines="1"
 
             android:singleLine="true"
 
-            android:textColor="@color/track_view_first_line_color"
-            android:textFontWeight="400"
-
-            android:textSize="@dimen/trackview_first_line_textsize"
             tools:text="@string/track_view_first_line_test_title" />
 
         <TextView
             android:id="@+id/artistName"
 
+            style="@style/TrackViewSecondLineStyle"
+
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
 
-            android:fontFamily="@font/ys_display_regular"
             android:maxLines="1"
 
             android:singleLine="true"
 
-            android:textColor="@color/track_view_second_line_color"
-            android:textFontWeight="400"
-            android:textSize="@dimen/trackview_second_line_textsize"
             tools:text="@string/track_view_second_line_test_title" />
     </LinearLayout>
