+++ app/src/main/res/values/styles.xml
@@ -80,11 +80,33 @@
         <item name="android:background">@color/white</item>
     </style>
 
-    <style name="SearchActivityEditTextColorStyle">
+    <style name="SearchActivityEditTextStyle">
         <item name="android:drawableTint">@color/grey</item>
         <item name="android:textColorHint">@color/grey</item>
         <item name="android:textColor">@color/black</item>
+
+        <item name="android:fontFamily">@font/ys_display_regular</item>
+        <item name="android:textFontWeight">400</item>
+        <item name="android:textSize">16sp</item>
     </style>
+
+    <!-- styles for track_view -->
+
+    <style name="TrackViewFirstLineStyle">
+        <item name="android:textColor">@color/track_view_first_line_color</item>
+        <item name="android:fontFamily">@font/ys_display_regular</item>
+        <item name="android:textFontWeight">400</item>
+        <item name="android:textSize">@dimen/trackview_first_line_textsize</item>
+    </style>
+
+    <style name="TrackViewSecondLineStyle">
+        <item name="android:textColor">@color/track_view_second_line_color</item>
+        <item name="android:fontFamily">@font/ys_display_regular</item>
+        <item name="android:textFontWeight">400</item>
+        <item name="android:textSize">@dimen/trackview_second_line_textsize</item>
+    </style>
+
+
 
 
 </resources>
